refactor: reorganize todo functionality by extracting helper functions for better modularity
